requirements.txt
PyQt5
opencv-python
numpy
Pillow